# api/openapi.yaml
openapi: 3.0.3
info:
  title: Helling API
  version: 0.2.0
  description: |
    Helling is a self-hosted, single-node container and VM management platform.
    This is the v0.2 Platform Core bridge: authentication, instance CRUD,
    operation polling, schedules, webhooks, host firewall, events, audit, and
    user Incus scope assignment.

    All endpoints under `/v1/` are stable within v0.2. Breaking changes ship under `/v2/`.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Bizarre Industries
    url: https://github.com/Bizarre-Industries/helling
servers:
  - url: https://localhost/v1
    description: Local installation behind the Caddy edge service
  - url: http+unix:///run/helling/api.sock/v1
    description: Direct socket access (CLI, dev)
...
security:
  - cookieAuth: []
paths:
...
  /webhooks:
    get:
      operationId: listWebhooks
      tags:
        - webhooks
      summary: List webhooks
      description: Returns webhook definitions without secrets. Admin role required.
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListEnvelope"
              example:
                data:
                  - id: 018f3a50-0000-7000-8000-000000000000
                    name: ops
                    url: https://hooks.example.com/helling
                    events:
                      - instance.started
                    enabled: true
                    created_at: "2026-05-05T09:00:00Z"
                    updated_at: "2026-05-05T09:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createWebhook
      tags:
        - webhooks
      summary: Create a webhook
      description: Stores an age-encrypted HMAC secret and validates the HTTPS destination. Admin role required.
      requestBody:
        required: true
        description: Webhook creation request. Secret is write-only.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookCreateRequest"
            example:
              name: ops
              url: https://hooks.example.com/helling
              events:
                - instance.started
              secret: "redacted-example"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEnvelope"
              example:
                data:
                  id: 018f3a50-0000-7000-8000-000000000000
                  name: ops
                  url: https://hooks.example.com/helling
                  events:
                    - instance.started
                  enabled: true
                  created_at: "2026-05-05T09:00:00Z"
                  updated_at: "2026-05-05T09:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /webhooks/{id}:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      operationId: getWebhook
      tags:
        - webhooks
      summary: Get a webhook
      description: Returns one webhook without its secret. Admin role required.
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEnvelope"
              example:
                data:
                  id: 018f3a50-0000-7000-8000-000000000000
                  name: ops
                  url: https://hooks.example.com/helling
                  events:
                    - instance.started
                  enabled: true
                  created_at: "2026-05-05T09:00:00Z"
                  updated_at: "2026-05-05T09:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateWebhook
      tags:
        - webhooks
      summary: Update a webhook
      description: Partially updates webhook metadata and optionally rotates the secret. Admin role required.
      requestBody:
        required: true
        description: Partial webhook update request. Secret is write-only.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUpdateRequest"
            example:
              enabled: false
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEnvelope"
              example:
                data:
                  id: 018f3a50-0000-7000-8000-000000000000
                  name: ops
                  url: https://hooks.example.com/helling
                  events:
                    - instance.started
                  enabled: false
                  created_at: "2026-05-05T09:00:00Z"
                  updated_at: "2026-05-05T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteWebhook
      tags:
        - webhooks
      summary: Delete a webhook
      description: Deletes a webhook and delivery history. Admin role required.
      responses:
        "204":
          description: Webhook deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: helling_session
      description: Session cookie issued by `POST /auth/login`. Sent automatically by the browser; CLI stores it under `~/.config/helling/session`.
  parameters:
...
    UUID:
      in: path
      name: id
      required: true
      description: UUID path parameter.
      schema:
        type: string
        description: UUID v7 string.
        format: uuid
        example: 018f3a40-0000-7000-8000-000000000000
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated principal lacks the required role or token scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Webhook:
      type: object
      description: Outbound webhook definition without secret material.
      required: [id, name, url, events, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          description: Webhook UUID.
          format: uuid
          example: 018f3a50-0000-7000-8000-000000000000
        name:
          type: string
          description: Human label for the webhook.
          minLength: 1
          maxLength: 63
          example: ops
        url:
          type: string
          description: HTTPS webhook destination.
          format: uri
          example: https://hooks.example.com/helling
        events:
          type: array
          description: Event types delivered to this webhook.
          items:
            type: string
            description: Helling event type.
            example: instance.started
        enabled:
          type: boolean
          description: True when delivery is enabled.
          example: true
        created_at:
          type: string
          description: Webhook creation time.
          format: date-time
          example: "2026-05-05T09:00:00Z"
        updated_at:
          type: string
          description: Webhook update time.
          format: date-time
          example: "2026-05-05T09:00:00Z"
    WebhookCreateRequest:
      type: object
      description: Request body for creating a webhook.
      required: [name, url, events, secret]
      properties:
        name:
          type: string
          description: Human label for the webhook.
          minLength: 1
          maxLength: 63
          example: ops
        url:
          type: string
          description: HTTPS webhook destination.
          format: uri
          example: https://hooks.example.com/helling
        events:
          type: array
          description: Event types delivered to this webhook.
          minItems: 1
          example: [instance.started]
          items:
            type: string
            description: Helling event type.
            example: instance.started
        secret:
          type: string
          description: HMAC secret. Stored encrypted and never returned.
          minLength: 16
          maxLength: 512
          format: password
          writeOnly: true
          example: redacted-example
        enabled:
          type: boolean
          description: Initial delivery enabled state.
          default: true
          example: true
    WebhookUpdateRequest:
      type: object
      description: Partial request body for updating a webhook.
      properties:
        name:
          type: string
          description: Replacement human label.
          minLength: 1
          maxLength: 63
          example: ops
        url:
          type: string
          description: Replacement HTTPS webhook destination.
          format: uri
          example: https://hooks.example.com/helling
        events:
          type: array
          description: Replacement event type list.
          example: [instance.started]
          items:
            type: string
            description: Helling event type.
            example: instance.started
        secret:
          type: string
          description: Replacement HMAC secret. Stored encrypted and never returned.
          minLength: 16
          maxLength: 512
          format: password
          writeOnly: true
          example: redacted-example
        enabled:
          type: boolean
          description: Replacement delivery enabled state.
          example: true
    WebhookEnvelope:
      type: object
      description: Envelope containing one webhook.
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Webhook"
    WebhookListEnvelope:
      type: object
      description: Envelope containing webhooks.
      required: [data]
      properties:
        data:
          type: array
          description: Webhook payload list.
          example:
            - id: 018f3a50-0000-7000-8000-000000000000
              name: ops
              url: https://hooks.example.com/helling
              events:
                - instance.started
              enabled: true
              created_at: "2026-05-05T09:00:00Z"
              updated_at: "2026-05-05T09:00:00Z"
          items:
            $ref: "#/components/schemas/Webhook"
...
    Error:
      type: object
      description: Standard error envelope returned by 4xx and 5xx responses.
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable error code.
          example: not_found
        message:
          type: string
          description: Human-readable message.
          example: instance not found
        details:
          type: object
          description: Optional structured context.
          additionalProperties: true
          example:
            field: name